# docs/swagger.yaml
basePath: /
definitions:
  ant_internal_platform_render.Response:
    properties:
      data: {}
      error:
        type: string
      status:
        type: integer
    type: object
...
  internal_order.CreateOrderRequest:
    properties:
      customer_contact:
        maxLength: 20
        minLength: 7
        type: string
      customer_name:
        maxLength: 100
        type: string
      device_id:
        description: |-
          DeviceID identifies the client device for returning-customer recognition
          on the public order-intake page. Optional; the IP is captured server-side.
        maxLength: 64
        type: string
      group_id:
        description: |-
          GroupID attaches the order to an existing group (tab). When omitted, a
          new group is minted in the same transaction and attached automatically;
          its token is returned in the order's group_token so the UI can reuse it
          to attach later orders to the same tab.
        minimum: 1
        type: integer
      group_label:
        description: |-
          GroupLabel optionally names the tab when a new group is auto-created
          (ignored when group_id is supplied).
        maxLength: 100
        type: string
      ordered_at:
        description: OrderedAt sets the business order date; defaults to now when
          omitted.
        type: string
      products:
        items:
          $ref: '#/definitions/internal_order.OrderItemRequest'
        minItems: 1
        type: array
      status:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      tax_percent:
        description: |-
          TaxPercent is the order tax rate as a percentage (0–100). Defaults to 0
          when omitted.
        maximum: 100
        minimum: 0
        type: number
    required:
    - customer_contact
    - customer_name
    - products
    type: object
  internal_order.CreatePublicOrderRequest:
    properties:
      customer_contact:
        maxLength: 20
        minLength: 7
        type: string
      customer_name:
        maxLength: 100
        type: string
      device_id:
        description: |-
          DeviceID identifies the client device for returning-customer recognition
          on the public order-intake page. Optional; the IP is captured server-side.
        maxLength: 64
        type: string
      group_id:
        description: |-
          GroupID attaches the order to an existing group (tab). When omitted, a
          new group is minted in the same transaction and attached automatically;
          its token is returned in the order's group_token so the UI can reuse it
          to attach later orders to the same tab.
        minimum: 1
        type: integer
      group_label:
        description: |-
          GroupLabel optionally names the tab when a new group is auto-created
          (ignored when group_id is supplied).
        maxLength: 100
        type: string
      ordered_at:
        description: OrderedAt sets the business order date; defaults to now when
          omitted.
        type: string
      products:
        items:
          $ref: '#/definitions/internal_order.OrderItemRequest'
        minItems: 1
        type: array
      status:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
      tax_percent:
        description: |-
          TaxPercent is the order tax rate as a percentage (0–100). Defaults to 0
          when omitted.
        maximum: 100
        minimum: 0
        type: number
      website:
        maxLength: 0
        type: string
    required:
    - customer_contact
    - customer_name
    - products
    type: object
  internal_order.Order:
    properties:
      app_id:
        type: integer
      created_at:
        type: string
      customer_contact:
        type: string
      customer_name:
        type: string
      device_id:
        description: |-
          DeviceID is the client-supplied device identifier used for returning-
          customer recognition on the public order-intake page.
        type: string
      division_id:
        type: integer
      group_id:
        type: integer
      group_token:
        type: string
      id:
        type: integer
      ip_address:
        description: IPAddress is the client IP captured server-side at creation.
          Read-only.
        type: string
      ordered_at:
        type: string
      products:
        items:
          $ref: '#/definitions/internal_order.OrderItem'
        type: array
      products_count:
        type: integer
      status:
        type: integer
      tax_percent:
        description: |-
          TaxPercent is the tax rate applied to the order as a percentage
          (e.g. 18.5 = 18.5%). Range 0–100.
        type: number
      total:
        description: |-
          Total is the pre-tax order amount: sum over items of (base price + chosen
          option deltas) * quantity. Persisted as a denormalized column and
          maintained on every item change, so it is populated on both list and
          detail reads.
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  internal_order.OrderItem:
    properties:
      attributes:
        items:
          $ref: '#/definitions/internal_order.OrderItemAttribute'
        type: array
      id:
        type: integer
      line_total:
        description: LineTotal is (Price + sum of attribute PriceDelta) * Quantity.
        type: number
      price:
        type: number
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
    type: object
  internal_order.OrderItemAttribute:
    properties:
      attribute_id:
        type: integer
      attribute_name:
        type: string
      option_id:
        type: integer
      option_value:
        type: string
      price_delta:
        type: number
    type: object
  internal_order.OrderItemAttributeRequest:
    properties:
      attribute_id:
        type: integer
      option_id:
        type: integer
    required:
    - attribute_id
    - option_id
    type: object
  internal_order.OrderItemRequest:
    properties:
      attributes:
        items:
          $ref: '#/definitions/internal_order.OrderItemAttributeRequest'
        type: array
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  internal_order.SetOrderGroupRequest:
    properties:
      group_id:
        minimum: 1
        type: integer
    required:
    - group_id
    type: object
  internal_order.SyncOrderItemRequest:
    properties:
      attributes:
        items:
          $ref: '#/definitions/internal_order.OrderItemAttributeRequest'
        type: array
      id:
        minimum: 1
        type: integer
      product_id:
        minimum: 1
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
  internal_order.UpdateOrderRequest:
    properties:
      customer_contact:
        maxLength: 20
        minLength: 7
        type: string
      customer_name:
        maxLength: 100
        type: string
      ordered_at:
        type: string
      products:
        items:
          $ref: '#/definitions/internal_order.SyncOrderItemRequest'
        minItems: 1
        type: array
      tax_percent:
        description: |-
          TaxPercent is the order tax rate as a percentage (0–100). When omitted the
          existing value is preserved.
        maximum: 100
        minimum: 0
        type: number
    required:
    - customer_contact
    - customer_name
    - products
    type: object
  internal_order.UpdateOrderStatusRequest:
    properties:
      status:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
    required:
    - status
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: This is the ant microservice.
  title: Ant API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Get a summary list of orders (no items, products_count only) for
        the authenticated app
      parameters:
      - description: Max items to return (default 50, max 500)
        in: query
        name: limit
        type: integer
      - description: Items to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by status (1=pending, 2=confirmed, 3=completed, 4=cancelled,
          5=paid)
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_order.Order'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: List all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create an order with one or more items. Items snapshot the product
        name, price and chosen attribute values; only active products can be added.
      parameters:
      - description: Order object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      description: Delete an order and all its items
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Delete order by ID
      tags:
      - orders
    get:
      description: Get a single order with its items
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Get order by ID
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Atomically update customer details and sync items in one call.
        Items with an id update quantity only (product/attributes immutable); items
        without an id are added; existing items missing from the payload are deleted.
        Status is managed via /orders/{id}/status.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Update order by ID
      tags:
      - orders
  /orders/{id}/group:
    patch:
      consumes:
      - application/json
      description: Set group_id to attach the order to an order group (tab), or omit/null
        to detach it.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group assignment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.SetOrderGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Attach/detach order to a group
      tags:
      - orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Set the order status (1=pending, 2=confirmed, 3=completed, 4=cancelled,
        5=paid)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Set order status
      tags:
      - orders
...
  /public/orders:
    post:
      consumes:
      - application/json
      description: Creates a pending order on behalf of a guest-token caller, optionally
        attaching it to an existing group (group_id) or auto-minting a new one. The
        order is always created pending — confirmation and payment are done at reception.
        Protected by reCAPTCHA, a hidden honeypot field, and a per-device volume cap.
      parameters:
      - description: Order object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.CreatePublicOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Create an order (public)
      tags:
      - Public
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...